feat: mc commit
consultas

+++ Domain/Entities/CursoEscolar.cs
@@ -6,9 +6,10 @@ namespace Domain.Entities;
 public partial class CursoEscolar : BaseEntity
 {    
 
-    public int AnioInicio { get; set; }
+    
+    public short AnioInicio { get; set; }
 
-    public int AnioFin { get; set; }
+    public short AnioFin { get; set; }
 
     public virtual ICollection<AlumnoSeMatriculaAsignatura> AlumnoSeMatriculaAsignaturas { get; set; } = new List<AlumnoSeMatriculaAsignatura>();
 }
